Fix running multiple processes on same core
We are currently not able to run two (non-init) processes/domains on the
same core (for example, nameserver and shell). Something goes wrong when
two processes do RPC-s at the same time. Specifically, whenever we call
lmp_chan_register_send() (because we are not able to send an LMP message
immediately because the other process is sending lots of messages), then
the send closure is never called and the thread remains stuck waiting
forever for the "done" bit.

I am not sure why. lmp_chan_register_send() puts the channel in
lmp_send_events_list, and it seems the channel needs to be triggered
from lmp_channels_retry_send_disabled(). But the latter is only called
if the process gets scheduled again (disp_run() is called). Currently
the process calls event_dispatch() which seems to put the process to
sleep until an event arrives (such as an LMP message), and until then
the process is not scheduled again. And therefore
lmp_channels_retry_send_disabled() is not called, and so the send
closure neither.

The issue seems fixed if the process instead calls
event_dispatch_non_block() and thread_yield(). Then it gets scheduled
again, and triggers the send closure.