Nameservice rpc implementation for LMP
This is a first version of the nameservice rpc function. The client
allocates a frame which acts as send buffer. Then the service name
and the message content are copied into the buffer. After that the
frame cap and the size of the buffer is sent over LMP to the server.

The server maps the frame into his address space and reads the service
name from the buffer. Then he consult the hash table to find the service
handler. He calls the service handler and collects the response in
a buffer (in a different frame). He sends the frame cap back to the
client the same way as described before.

The receivers free the resources (although frame_free doesn't work
somehow) but I'm not sure if this is the best approach. On the sender
side I just unmap and delete the frame capabilities. This needs some
testing especially if multiple domains accessing the same service.